Remove temporary BC compatibility codes
