+++ src/fairseq2/assets/card.py
@@ -151,19 +151,6 @@ class AssetCard:
     def __repr__(self) -> str:
         return repr(self._metadata)
 
-    # compat
-    def asset_type(self) -> str:
-        """Return the type of the asset represented by this card."""
-        for field in ["model_type", "dataset_type", "tokenizer_type"]:
-            try:
-                return self.field(field).as_(str)  # type: ignore[no-any-return]
-            except AssetCardFieldNotFoundError:
-                continue
-
-        raise AssetCardFieldNotFoundError(
-            f"The asset card '{self.name}' must have a field named 'model_type', 'dataset_type', or 'tokenizer_type'."
-        )
-
     @property
     def name(self) -> str:
         """The name of the asset."""
