+++ src/fairseq2/assets/cards/models/llama.yaml
@@ -5,7 +5,6 @@
 # LICENSE file in the root directory of this source tree.
 
 name: llama
-model_type: llama  # compat
 model_family: llama
 checkpoint: "https://ai.meta.com/llama/;gated=true"
 tokenizer: "https://ai.meta.com/llama/;gated=true"
@@ -20,7 +19,6 @@ model_arch: 7b
 ---
 
 name: llama2
-model_type: llama  # compat
 model_family: llama
 checkpoint: "https://ai.meta.com/llama/;gated=true"
 tokenizer: "https://ai.meta.com/llama/;gated=true"
@@ -67,7 +65,6 @@ num_shards: 8
 ---
 
 name: llama3
-model_type: llama  # compat
 model_family: llama
 checkpoint: "https://ai.meta.com/llama/;gated=true"
 tokenizer: "https://ai.meta.com/llama/;gated=true"
