+++ src/fairseq2/assets/cards/models/mistral.yaml
@@ -5,7 +5,6 @@
 # LICENSE file in the root directory of this source tree.
 
 name: mistral_7b
-model_type: mistral  # compat
 model_family: mistral
 model_arch: 7b
 checkpoint: "https://files.mistral-7b-v0-1.mistral.ai/mistral-7B-v0.1.tar;path=mistral-7B-v0.1%2Fconsolidated.00.pth"
@@ -15,7 +14,6 @@ tokenizer_family: mistral
 ---
 
 name: mistral_7b_instruct
-model_type: mistral  # compat
 model_family: mistral
 model_arch: 7b
 checkpoint: "https://files.mistral-7b-v0-1.mistral.ai/mistral-7B-instruct-v0.1b.tar;path=Mistral-7B-instruct-v0.1%2Fconsolidated.00.pth"
