+++ src/fairseq2/assets/cards/models/s2t_conformer.yaml
@@ -5,7 +5,6 @@
 # LICENSE file in the root directory of this source tree.
 
 name: s2t_conformer_covost_st_en_de
-model_type: s2t_transformer  # compat
 model_family: s2t_transformer
 model_arch: conformer_medium
 task: translation
@@ -17,7 +16,6 @@ tokenizer_family: s2t_transformer
 ---
 
 name: s2t_conformer_covost_st_en_de_rel_pos
-model_type: s2t_transformer  # compat
 model_family: s2t_transformer
 model_arch: conformer_medium
 model_config:
