+++ src/fairseq2/checkpoint.py
@@ -84,11 +84,6 @@ class CheckpointManager(ABC):
             The model to save.
         """
 
-    # compat
-    @abstractmethod
-    def save_consolidated_model(self, step_nr: int, model: Module) -> None:
-        ...
-
     @abstractmethod
     def load_checkpoint(self, step_nr: int) -> Dict[str, Any]:
         """Load the checkpoint of the specified training step.
@@ -150,15 +145,6 @@ class CheckpointManager(ABC):
     @abstractmethod
     def get_step_numbers(self) -> List[int]:
         """Return the numbers of the training steps that have a checkpoint."""
-
-    # compat
-    @abstractmethod
-    def get_last_step_number(self, *, with_model: bool = False) -> Optional[int]:
-        """Return the number of the last training step that has a checkpoint.
-
-        :param with_model:
-            If ``True``, only considers training steps with a saved model.
-        """
 
 
 @final
@@ -489,11 +475,6 @@ class FileCheckpointManager(CheckpointManager):
 
         self._root_gang.barrier()
 
-    # compat
-    @override
-    def save_consolidated_model(self, step_nr: int, model: Module) -> None:
-        self.save_consolidated_fsdp_model(step_nr, model)
-
     @override
     def load_checkpoint(self, step_nr: int) -> Dict[str, Any]:
         def raise_error(cause: Exception) -> NoReturn:
@@ -657,7 +638,11 @@ class FileCheckpointManager(CheckpointManager):
             the last model in the training.
         """
         if step_nr is None:
-            step_nr = self.get_last_step_number(with_model=True)
+            step_numbers = self.get_step_numbers()
+            if not step_numbers:
+                return None
+
+            step_nr = step_numbers[-1]
 
         if step_nr is None:
             return None
@@ -682,14 +667,6 @@ class FileCheckpointManager(CheckpointManager):
         step_numbers.sort()
 
         return step_numbers
-
-    # compat
-    @override
-    def get_last_step_number(self, *, with_model: bool = False) -> Optional[int]:
-        if step_numbers := self.get_step_numbers():
-            return step_numbers[-1]
-
-        return None
 
     def _iter_step_numbers(self) -> Iterator[int]:
         try:
