--- src/fairseq2/data/parquet_tools.py
@@ -1,3 +1,0 @@
-# compat
-from fairseq2.data.parquet.tools import *  # noqa: F401,F403
-from fairseq2.data.parquet.tools import _to_real_object  # noqa: F401,F403
