+++ src/fairseq2/logging.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import logging
 from logging import Logger, getLogger
-from typing import Any, Final, Optional, Set, final
+from typing import Any, Final, Optional, final
 
 
 @final
@@ -18,7 +18,6 @@ class LogWriter:
     _NO_HIGHLIGHT: Final = {"highlighter": None}
 
     _logger: Logger
-    _once_messages: Set[str]
 
     def __init__(self, logger: Logger) -> None:
         """
@@ -27,24 +26,11 @@ class LogWriter:
         """
         self._logger = logger
 
-        self._once_messages = set()
-
     def debug(
         self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
     ) -> None:
         """Log a message with level ``DEBUG``."""
         self._write(logging.DEBUG, msg, args, kwargs, highlight)
-
-    def debug_once(
-        self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
-    ) -> None:
-        """Log a message only once with level ``DEBUG``."""
-        if msg in self._once_messages:
-            return
-
-        self._write(logging.DEBUG, msg, args, kwargs, highlight)
-
-        self._once_messages.add(msg)
 
     def info(
         self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
@@ -52,50 +38,17 @@ class LogWriter:
         """Log a message with level ``INFO``."""
         self._write(logging.INFO, msg, args, kwargs, highlight)
 
-    def info_once(
-        self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
-    ) -> None:
-        """Log a message only once with level ``INFO``."""
-        if msg in self._once_messages:
-            return
-
-        self._write(logging.INFO, msg, args, kwargs, highlight)
-
-        self._once_messages.add(msg)
-
     def warning(
         self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
     ) -> None:
         """Log a message with level ``WARNING``."""
         self._write(logging.WARNING, msg, args, kwargs, highlight)
 
-    def warning_once(
-        self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
-    ) -> None:
-        """Log a message only once with level ``WARNING``."""
-        if msg in self._once_messages:
-            return
-
-        self._write(logging.WARNING, msg, args, kwargs, highlight)
-
-        self._once_messages.add(msg)
-
     def error(
         self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
     ) -> None:
         """Log a message with level ``ERROR``."""
         self._write(logging.ERROR, msg, args, kwargs, highlight)
-
-    def error_once(
-        self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
-    ) -> None:
-        """Log a message only once with level ``ERROR``."""
-        if msg in self._once_messages:
-            return
-
-        self._write(logging.ERROR, msg, args, kwargs, highlight)
-
-        self._once_messages.add(msg)
 
     def exception(
         self, msg: Any, *args: Any, highlight: bool = False, **kwargs: Any
