--- src/fairseq2/models/utils/__init__.py
@@ -1,16 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from __future__ import annotations
-
-from fairseq2.models.utils.arch_registry import (
-    ArchitectureRegistry as ArchitectureRegistry,
-)
-from fairseq2.models.utils.generic_loaders import ConfigLoader as ConfigLoader
-from fairseq2.models.utils.generic_loaders import ModelLoader as ModelLoader
-from fairseq2.models.utils.generic_loaders import (
-    StandardTextTokenizerLoader as StandardTextTokenizerLoader,
-)
