+++ src/fairseq2/nn/__init__.py
@@ -13,7 +13,6 @@ from fairseq2.nn.embedding import VocabShardedEmbedding as VocabShardedEmbedding
 from fairseq2.nn.embedding import init_scaled_embedding as init_scaled_embedding
 from fairseq2.nn.incremental_state import IncrementalState as IncrementalState
 from fairseq2.nn.incremental_state import IncrementalStateBag as IncrementalStateBag
-from fairseq2.nn.module_list import ModuleList as ModuleList
 from fairseq2.nn.normalization import LayerNorm as LayerNorm
 from fairseq2.nn.normalization import RMSNorm as RMSNorm
 from fairseq2.nn.normalization import StandardLayerNorm as StandardLayerNorm
