--- src/fairseq2/utils/logging.py
@@ -1,3 +1,0 @@
-# compat
-from fairseq2.logging import *  # noqa: F401,F403
-from fairseq2.recipes.logging import *  # noqa: F401,F403
