+++ tests/integration/generation/test_sampling.py
@@ -9,7 +9,8 @@ from typing import Final
 import torch
 
 from fairseq2.generation import SamplingSeq2SeqGenerator, TextTranslator, TopKSampler
-from fairseq2.models.nllb import load_nllb_model, load_nllb_tokenizer
+from fairseq2.models.nllb import load_nllb_tokenizer
+from fairseq2.models.transformer import load_transformer_model
 from tests.common import device
 
 ENG_SENTENCE1: Final = "On Monday, scientists from the Stanford University School of Medicine announced the invention of a new diagnostic tool that can sort cells by type: a tiny printable chip that can be manufactured using standard inkjet printers for possibly about one U.S. cent each."
@@ -21,7 +22,7 @@ DEU_SENTENCE2: Final = "Wie geht es Ihnen heute?"
 def test_greedy_sampling() -> None:
     model_name = "nllb-200_dense_distill_600m"
 
-    model = load_nllb_model(
+    model = load_transformer_model(
         model_name, device=device, dtype=torch.float32, progress=False
     )
